+++ crates/chain-config/Cargo.toml
@@ -26,10 +26,6 @@ serde = { workspace = true, features = ["derive", "rc"] }
 serde_json = { version = "1.0", features = ["raw_value"], optional = true }
 serde_with = { workspace = true, features = ["macros", "hex"] }
 tracing = { version = "0.1", default-features = false, optional = true }
-uuid = { workspace = true }
-tokio = { workspace = true, features = ["time"] }
-tokio-util = { workspace = true }
-futures = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt"] }
