+++ crates/chain-config/src/lib.rs
@@ -7,12 +7,10 @@ pub mod config;
 pub mod fee_collection_contract;
 mod genesis;
 mod serialization;
-mod task_manager;
 
 pub use config::*;
 use fuel_core_types::fuel_vm::SecretKey;
 pub use genesis::GenesisCommitment;
-pub use task_manager::*;
 
 /// A default secret key to use for testing purposes only
 pub fn default_consensus_dev_key() -> SecretKey {
